apps/backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0

python-jose [cryptography]
passlib[bcrypt]
PyJWT
python-decouple
python-multipart
sqlalchemy
databases
openai
requests